# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API RESTful - Concesionario de Autos",
        "version": "1.0.0",
        "description": "Este proyecto es una API RESTful desarrollada con Node.js y Express. Permite realizar operaciones CRUD sobre dos entidades: una principal y una de soporte. El proyecto incluye autenticación, validación de datos y manejo de errores centralizado."
    },
    "servers": [
        {
            "url": "https://concesionario-back.onrender.com/",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:3000",
            "description": "Local Server"
        }
    ],
    "components": {
        "schemas": {
...
            "Seller": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "The auto-generated id of the seller.",
                        "format": "objectId",
                        "example": "67436c0ecc80bce34f5274d0"
                    },
                    "lastname": {
                        "type": "string",
                        "description": "The lastname of the seller.",
                        "example": "Doe"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the seller.",
                        "example": "John"
                    },
                    "adress": {
                        "type": "string",
                        "description": "The address of the seller.",                        
                        "example": "Street 123"
                    },
                    "postcode": {
                        "type": "string",
                        "description": "The postcode of the seller.",
                        "example": "12345"
                    },
                    "city": {
                        "type": "string",
                        "description": "The city of the seller.",
                        "example": "City"
                    },
                    "province": {
                        "type": "string",
                        "description": "The province of the seller.",
                        "example": "Province"
                    },
                    "country": {
                        "type": "string",
                        "description": "The country of the seller.",
                        "example": "Country"
                    },
                    "phone": {
                        "type": "string",
                        "description": "The phone of the seller.",
                        "example": "123456789"
                    },
                    "cellphone": {
                        "type": "string",
                        "description": "The cellphone of the seller.",
                        "example": "123456789"
                    },
                    "email": {
                        "type": "string",
                        "description": "The email of the seller.",
                        "example": "mail@example.com"
                    }
                }
            }
        },
        "responses": {
            "400": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "car found"
                                },
                                "data": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            },    
            "500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "failed"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "server error"
                                },
                                "data": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "paths": {
...
        "/sellers": {
            "get": {
                "summary": "Gets a list of sellers.",
                "tags": ["Sellers"],
                "operationId": "getSellers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "sellers found"
                                        },
                                        "data": {
                                            "type": "array",
                                            "$ref": "#/components/schemas/Seller"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500":{"$ref": "#/components/responses/500"}
                }
            },
            "post": {
                "summary": "Creates a new seller.",
                "tags": ["Sellers"],
                "operationId": "createSeller",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Seller"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "seller created"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Seller"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400":{"$ref": "#/components/responses/400"},
                    "500":{"$ref": "#/components/responses/500"}
                }
            }    
        },
        "/sellers/{id}": {
            "get": {
                "summary": "Gets a seller by id.",
                "tags": ["Sellers"],
                "operationId": "getSeller",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the seller.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "objectId",
                            "example": "67436c0ecc80bce34f5274d0"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "seller found"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Seller"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400":{"$ref": "#/components/responses/400"},    
                    "500":{"$ref": "#/components/responses/500"}
                }
            },
            "put": {
                "summary": "Updates a seller by id.",
                "tags": ["Sellers"],
                "operationId": "updateSeller",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the seller.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "objectId",
                            "example": "67436c0ecc80bce34f5274d0"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Seller"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "seller updated"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Seller"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400":{"$ref": "#/components/responses/400"},    
                    "500":{"$ref": "#/components/responses/500"}
                }
            },
            "delete": {
                "summary": "Deletes a seller by id.",
                "tags": ["Sellers"],
                "operationId": "deleteSeller",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the seller.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "objectId",
                            "example": "67436c0ecc80bce34f5274d0"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "seller deleted"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Seller"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400":{"$ref": "#/components/responses/400"},    
                    "500":{"$ref": "#/components/responses/500"}
                }
            }
        }
    }
}